requirements.txt
django==5.1.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
python-dotenv==1.0.1